// src/types/openEvent/EventBuilderError.ts
export interface EventBuilderError {
  field: string | null;
  title: string | null;
  applicationPeriod: string | null;
  eventDates: string | null;
  address: string | null;
  cost: string | null;
  recruitmentNumber: string | null;
  description: string | null;
  imageUrls: string | null;
  additionalInformation: string | null;
  hostName: string | null;
  hostPhoneNumber: string | null;
  hostEmail: string | null;
}